add docs and verify

+++ fps/nth-term.hpp
@@ -6,13 +6,13 @@ using namespace std;
 #include "kitamasa.hpp"
 
 template <typename mint>
-mint nth_term(int n, const vector<mint> &s) {
+mint nth_term(long long n, const vector<mint> &s) {
   using fps = FormalPowerSeries<mint>;
   auto bm = BerlekampMassey<mint>(s);
   return kitamasa(n, fps{begin(bm), end(bm)}, fps{begin(s), end(s)});
 }
 
 /**
- * @brief 与えられた数列の第$n$項を計算(Berlekamp-Massey/Bostan-Mori)
- * @docs docs/fps/kitamasa.md
+ * @brief 線形回帰数列の高速計算(Berlekamp-Massey/Bostan-Mori)
+ * @docs docs/fps/nth-term.md
  */
